+++ src/duplocloud/args.py
@@ -303,16 +303,22 @@ TIME = Arg("time", "--time",
 TITLE = Arg("title", "--title",
             help= "The Title for ticket")
 
-AGENTNAME = Arg("agent_name", "--agent_name",
+AGENTNAME = Arg("agent_name", "--agent_name", "--agent",
                 help= "AI Agent to be used to process the ticket")
 
-INSTANCEID = Arg("instance_id","--instance_id",
+INSTANCEID = Arg("instance_id","--instance_id", "--instance",
                 help= "AI Agent Instance Id")
 
 APIVERSION = Arg("api_version", "--api-version",
                  help="API Version",
                  required=False,
                  default="v1")
+
+TICKETID = Arg("ticket_id", "--ticket_id", "--ticket",
+               help="The ID of the AI HelpDesk ticket")
+
+Message = Arg("content", "--content", "--msg",
+              help="The message you want to send to the AI agent")
 
 DAYS = Arg("days", 
             help='The days to use',